half and mixed precision inference (#442)
* implement half and mixed precision inference in Pipeline

* cast probabiliteis to float before converting to numpy

* add test cases for parameters half_precision_model and half_precision_ops to test_re_text_classification

* add docs for new params

* fix spellings in docs

* fix documentation

* fix documentation

* add todo

* move get_autocast_dtype out of Pipeline class

+++ src/pytorch_ie/taskmodules/transformer_text_classification.py
@@ -196,7 +196,7 @@ class TransformerTextClassificationTaskModule(TaskModuleType):
         logits = model_output["logits"]
 
         output_label_probs = logits.sigmoid() if self.multi_label else logits.softmax(dim=-1)
-        output_label_probs = output_label_probs.detach().cpu().numpy()
+        output_label_probs = output_label_probs.detach().cpu().float().numpy()
 
         if self.multi_label:
             raise NotImplementedError()
